REFACTOR steps/* + TESTS PASSED
