+++ src/plugins/plugins.md
@@ -5,19 +5,19 @@
 
 # plugins/
     main/
-        home
-        footer
-        _layout
-        404
+        ✔ home
+        ✔ footer
+        ✔ _layout
+        ✔ 404
     steps/
-        1-prijmy
-        2-zamestnanie
-        3-manzelstvo
-        3b-deti
-        3c-hypoteka
-        3d-dve-percenta
-        4-dochodok
-        5-osobne-udaje
+        ✔ 1-prijmy
+        ✔ 2-zamestnanie
+        ✔ 3-manzelstvo
+        ✔ 3b-deti
+        ✔ 3c-hypoteka
+        ✔ 3d-dve-percenta
+        ✔ 4-dochodok
+        ✔ 5-osobne-udaje
     summary/
         6-suhrn
         7-vratenie-preplatku
@@ -25,12 +25,12 @@
         9-pokracovat
         calculation
     odklad/
-        1-prijmy
-        2-osobne-udaje
-        3-suhrn
-        4-pokracovat
+        ✔ 1-prijmy
+        ✔ 2-osobne-udaje
+        ✔ 3-suhrn
+        ✔ 4-pokracovat
     _components/
-        form/
-        links/
+        ✔ form/
+        ✔ links/
     zz_unknown/
